SignalR, refactor

+++ src/Backend/UnderseaBackend/Undersea.BLL/Services/GameService.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using System.Threading.Tasks;
+using Undersea.BLL.Hubs;
 using Undersea.BLL.Interfaces;
 using Undersea.DAL;
 using Undersea.DAL.Models;
@@ -15,11 +16,7 @@ namespace Undersea.BLL.Services
         {
             get
             {
-                return context.Game.First().CurrentTurn;
-            }
-
-            set
-            {
+                return _context.Game.First().CurrentTurn;
             }
         }
 
@@ -28,18 +25,20 @@ namespace Undersea.BLL.Services
         private readonly ICityRepository _cityRepository;
         private readonly IArmyRepository _armyRepository;
         private readonly IAttackRepository _attackRepository;
-        private readonly AppDbContext context;
+        private readonly ISignalHub _signalHub;
+
+        private readonly AppDbContext _context;
 
         public GameService(IUserRepository userRepository, ICityRepository cityRepository, IArmyRepository armyRepository,
-                            IAttackRepository attackRepository, AppDbContext context, IArmyService armyService, ICityService cityService)
+                            IAttackRepository attackRepository, AppDbContext context, IArmyService armyService, ICityService cityService, ISignalHub signalHub)
         {
-            CurrentTurn = 1;
             _cityRepository = cityRepository;
             _armyRepository = armyRepository;
             _attackRepository = attackRepository;
-            this.context = context;
+            _context = context;
             _armyService = armyService;
             _cityService = cityService;
+            _signalHub = signalHub;
         }
 
         public GameService()
@@ -62,43 +61,48 @@ namespace Undersea.BLL.Services
             }
             var attacks = await _attackRepository.GetAll();
 
-            foreach (Attack a in attacks)
+            if (attacks.Any())
             {
-                int defense = await _armyService.GetArmyDefensePower(a.DefenderCity.AvailableArmyId);
-                int attack = await _armyService.GetArmyAttackingPower(a.ArmyId);
 
-                // TODO támadóerő +- 5%
-
-                if (attack > defense)
+                foreach (Attack a in attacks)
                 {
-                    a.AttackerCity.PearlCount += a.DefenderCity.PearlCount / 2;
-                    a.AttackerCity.CoralCount += a.DefenderCity.CoralCount / 2;
+                    int defense = await _armyService.GetArmyDefensePower(a.DefenderCity.AvailableArmyId);
+                    int attack = await _armyService.GetArmyAttackingPower(a.ArmyId);
 
-                    a.DefenderCity.CoralCount /= 2;
-                    a.DefenderCity.PearlCount /= 2;
+                    double moral = new Random().NextDouble() * 0.1;
+                    double moralAttack = (0.95 + moral) * attack; 
 
-                    foreach (ArmyUnit au in a.DefenderCity.AvailableArmy.Units)
+                    if (moralAttack > defense)
                     {
-                        au.UnitCount = Convert.ToInt32(Math.Floor(au.UnitCount * 0.9));
-                    }
-                }
+                        a.AttackerCity.PearlCount += a.DefenderCity.PearlCount / 2;
+                        a.AttackerCity.CoralCount += a.DefenderCity.CoralCount / 2;
 
-                else
-                {
+                        a.DefenderCity.CoralCount /= 2;
+                        a.DefenderCity.PearlCount /= 2;
+
+                        foreach (ArmyUnit au in a.DefenderCity.AvailableArmy.Units)
+                        {
+                            au.UnitCount = Convert.ToInt32(Math.Floor(au.UnitCount * 0.9));
+                        }
+                    }
+
+                    else
+                    {
+                        foreach (ArmyUnit au in a.Army.Units)
+                        {
+                            au.UnitCount = Convert.ToInt32(Math.Floor(au.UnitCount * 0.9));
+                        }
+                    }
+
                     foreach (ArmyUnit au in a.Army.Units)
                     {
-                        au.UnitCount = Convert.ToInt32(Math.Floor(au.UnitCount * 0.9));
+                        a.AttackerCity.AvailableArmy.Units.First(a => a.UnitType == au.UnitType).UnitCount += au.UnitCount;
                     }
-                }
 
-                foreach (ArmyUnit au in a.Army.Units)
-                {
-                    a.AttackerCity.AvailableArmy.Units.First(a => a.UnitType == au.UnitType).UnitCount += au.UnitCount;
+                    await _attackRepository.Remove(a);
+                    await _cityRepository.Update(a.DefenderCity);
+                    await _cityRepository.Update(a.AttackerCity);
                 }
-
-                await _cityRepository.Update(a.DefenderCity);
-                await _cityRepository.Update(a.AttackerCity);
-                await _attackRepository.Remove(a);
             }
 
             foreach (City c in cities)
@@ -107,10 +111,10 @@ namespace Undersea.BLL.Services
                 await _cityRepository.Update(c);
             }
 
-            //rank számítás
+            _context.Game.First().CurrentTurn++;
+            await _context.SaveChangesAsync();
 
-            context.Game.First().CurrentTurn++;
-            await context.SaveChangesAsync();
+            await _signalHub.SendMessage("Server", "Next turn");
         }
     }
 }
